Split dataset fixtures between conftest.py and fixtures
This makes it easier to stay on top of F811 flake8 errors as we can put
all pytest fixtures in conftest and then we can avoid importing and
redefining them.
